rag_requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
sentence-transformers
faiss-cpu